+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
 v0.89.2
 
 - Protocol: Update the `spqr` dependency to v1.5.1
+
+- Expose getUploadForm() to clients (for attachments)
+